Working state code update
In this commit I have made all the changes that make this app work in Kubernetes. This, if cloned into gcloud, dockerised and you apply the deployment and service files (having changed both to fit with your image and external static IP (or make an ingress if you want to forward to a domain name)) will work... with one catch. each user needs a individual pod because session management has not been implemented.